requirements.txt
easyocr
mss
Pillow
requests
pynput
pyautogui
